requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
scikit-learn>=1.6.0
imbalanced-learn>=0.13.0
matplotlib==3.9.0
seaborn==0.13.2
shap==0.45.0
xgboost==2.0.3
lightgbm==4.4.0
torch==2.3.0
torchvision==0.18.0
reportlab==4.2.0
optuna==3.6.1
groq>=0.9.0
python-dotenv>=1.0.0
